split app.tsx to app, footer, locationrouter
