+++ src/App.tsx
@@ -4,99 +4,15 @@ import "./App.css";
 import { CalendarSection } from "./calendar/CalendarSection";
 import LanguageSelect from "./LanguageSelect";
 import { LocationPicker } from "./location-picker/LocationPicker";
-import { ShareButtons } from "./ShareButtons";
-import { enqueueAnalyticsEvent } from "./utils/analytics";
 import { DEFAULT_LOCATION } from "./utils/consts";
 import { useSearchParams } from "./utils/url";
 import { TodayLocationsSection } from "./today-locations/TodayLocationsSection";
 import CookiesBar from "./Cookies";
 import BookingModal from "./calendar/modal/CalendarModal";
-import WalkModal from "./today-locations/healthpoint/HealthpointModal";
-import {
-  getHealthpointData,
-  HealthpointLocation,
-} from "./today-locations/healthpoint/HealthpointData";
-import { CrowdsourcedLocation } from "./crowdsourced/CrowdsourcedData";
-import CrowdsourcedModal from "./crowdsourced/CrowdsourcedModal";
-import {
-  Switch,
-  Route,
-  useParams,
-  useHistory,
-  useLocation,
-} from "react-router-dom";
+import { Switch, Route, useLocation } from "react-router-dom";
 import { useBookingData } from "./calendar/booking/BookingData";
-import { simpleHash } from "./utils/simpleHash";
-import { crowdsourcedLocations } from "./crowdsourced/CrowdsourcedLocations";
-
-function LocationRouter({ radiusKm }: { radiusKm: number }) {
-  const { slug } = useParams<{ slug: string }>();
-  const [hash] = slug.split("-").slice(-1);
-  const history = useHistory();
-
-  const [healthpointLocations, setHealthpointLocations] = useState<
-    HealthpointLocation[]
-  >([]);
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState<Error | null>(null);
-
-  useEffect(() => {
-    async function load() {
-      setLoading(true);
-      setError(null);
-      try {
-        setHealthpointLocations(await getHealthpointData());
-      } catch (e) {
-        setError(e as Error);
-      }
-      setLoading(false);
-    }
-    load();
-  }, []);
-
-  const locations = [...healthpointLocations, ...crowdsourcedLocations];
-
-  if (loading) {
-    return <p>Loading...</p>;
-  }
-
-  if (error) {
-    return <p>Error: {error.message}</p>;
-  }
-
-  if (!locations.length) {
-    return <p>No locations found</p>;
-  }
-
-  const location = locations.find(
-    (loc) => simpleHash(`${loc.lat}${loc.lng}`) === hash
-  );
-
-  if (!location) {
-    return (
-      <p>
-        {slug}|{hash}
-      </p>
-    );
-  }
-
-  return (
-    <>
-      {"isHealthpoint" in location ? (
-        <WalkModal
-          clearSelectedLocation={() => history.push("/")}
-          radiusKm={radiusKm}
-          location={location as HealthpointLocation}
-        />
-      ) : (
-        <CrowdsourcedModal
-          clearSelectedLocation={() => history.push("/")}
-          location={location as CrowdsourcedLocation}
-        />
-      )}
-    </>
-  );
-}
+import { LocationRouter } from "./LocationRouter";
+import { Footer } from "./Footer";
 
 function App() {
   const { lat, lng, radius } = useSearchParams();
@@ -216,76 +132,7 @@ function App() {
             </Route>
           </Switch>
         </div>
-        <footer className="footer-header">
-          <p style={{ marginBottom: "0.5rem" }}>{t("footer.message")}</p>
-          <div className={"social-container"}>
-            <ShareButtons />
-          </div>
-          <br />
-          <p>
-            <Trans
-              i18nKey="footer.addressFinderLinkCopy"
-              t={t}
-              // eslint-disable-next-line jsx-a11y/anchor-has-content
-              components={[<a href="https://addressfinder.nz"></a>]}
-            />
-          </p>
-          <br />
-          <p>
-            <a
-              href="https://airtable.com/shrxuw3vSp2yRPrG7"
-              target="_blank"
-              rel="noreferrer"
-            >
-              {t("footer.links.contactUs")}
-            </a>{" "}
-            /{" "}
-            <a
-              href="https://github.com/CovidEngine/vaxxnz/projects/2"
-              target="_blank"
-              rel="noreferrer"
-              onClick={() => enqueueAnalyticsEvent("Roadmap clicked")}
-            >
-              {t("footer.links.roadmap")}
-            </a>{" "}
-            /{" "}
-            <a
-              href="https://github.com/CovidEngine/vaxxnz/blob/main/CONTRIBUTING.md"
-              target="_blank"
-              rel="noreferrer"
-            >
-              {t("navigation.getInvolved")}
-            </a>
-            <br />{" "}
-            <a
-              href="https://github.com/CovidEngine/vaxxnzlocations"
-              target="_blank"
-              rel="noreferrer"
-              onClick={() => enqueueAnalyticsEvent("Raw data clicked")}
-            >
-              {t("footer.links.rawData")}
-            </a>{" "}
-            /{" "}
-            <a
-              href="https://github.com/CovidEngine/vaxxnz"
-              target="_blank"
-              rel="noreferrer"
-              onClick={() => enqueueAnalyticsEvent("Source code clicked")}
-            >
-              {t("footer.links.sourceCode")}
-            </a>{" "}
-          </p>
-          <p>
-            <a
-              href="https://twitter.com/vaxxnz"
-              target="_blank"
-              rel="noreferrer"
-              onClick={() => enqueueAnalyticsEvent("Twitter Clicked")}
-            >
-              Twitter
-            </a>{" "}
-          </p>
-        </footer>
+        <Footer />
       </div>
       <div className="background">
         <div
